+++ docs/versionhistory.rst
@@ -4,7 +4,7 @@ Version history
 To find out how to migrate your application from a previous version of
 APScheduler, see the :doc:`migration section <migration>`.
 
-**UNRELEASED**
+**4.0.0a3**
 
 - **BREAKING** The scheduler classes were moved to be importable (only) directly from
   the ``apscheduler`` package (``apscheduler.Scheduler`` and